# [chore] Bump fastapi version to speed up get_app