--- a/pyproject.toml
+++ b/pyproject.toml
@@ -108,7 +108,7 @@ celerybeat-mongo = { version = "^0.2.0", optional = true }
 cryptography = "^41.0.3"
 databricks-cli = { version = "^0.17.3", optional = true }
 databricks-sql-connector = { version = "2.8.0", optional = true }  # Awaiting snowflake-connector-python to support urllib3>=2.0
-fastapi = { version =  "^0.95.1", optional = true }
+fastapi = { version =  "^0.96.0", optional = true }
 featurebyte-freeware = { version = "^0.2.16", optional = true }
 gevent = {version = "^22.10.2", optional = true}
 humanize = "^4.4.0"