--- a/tests/unit/test_app.py
+++ b/tests/unit/test_app.py
@@ -34,7 +34,7 @@ def test_get_app__loading_time():
     start = time.time()
     get_app()
     elapsed_time = time.time() - start
-    print("get_app took %s seconds" % elapsed_time)
+    print(f"get_app took {elapsed_time} seconds")
     assert elapsed_time < 5