## Description

This bumps fastapi version to 0.96.0 to pick up https://github.com/tiangolo/fastapi/pull/4645 to see if there is any improvements in runtime.

Testing locally, `get_app()` timing in `test_get_app__loading_time` is around 10X faster (from around 8s to 0.8s).

Before:
```
============================= slowest 50 durations =============================
31.02s call     tests/unit/test_app.py::test_get_app__loading_time
25.93s call     tests/unit/api/test_feature.py::test_list_unsaved_features
21.03s call     tests/unit/api/test_feature_list.py::test_deploy
16.71s call     tests/unit/api/test_feature_list.py::test_save_feature_group
15.73s call     tests/unit/api/test_version.py::test_feature_and_feature_list_version
```

After:
```
============================= slowest 50 durations =============================
26.03s call     tests/unit/api/test_feature.py::test_list_unsaved_features
21.32s call     tests/unit/api/test_feature_list.py::test_deploy
16.76s call     tests/unit/api/test_feature_list.py::test_save_feature_group
15.67s call     tests/unit/api/test_user_defined_function.py::test_create_complex_feature_with_user_defined_function
15.46s call     tests/unit/api/test_version.py::test_feature_and_feature_list_version
```

## Related Issue

<!-- If your PR refers to a related issue, link it here. -->

## Type of Change

<!-- Mark with an `x` all the checkboxes that apply (like `[x]`) -->

Label this pull request to place it under the correct category in Release Notes:

|        **Pull Request Label**         | **Category in Release Notes** |
|:-------------------------------------:|:-----------------------------:|
|       `enhancement`, `feature`        |          🚀 Features          |
| `bug`, `refactoring`, `bugfix`, `fix` |    🔧 Fixes & Refactoring     |
|       `build`, `ci`, `testing`        |    📦 Build System & CI/CD    |
|              `breaking`               |      💥 Breaking Changes      |
|            `documentation`            |       📝 Documentation        |
|            `dependencies`             |    ⬆️ Dependencies updates    |



## Checklist

<!-- Mark with an `x` all the checkboxes that apply (like `[x]`) -->

- [ ] I have read the [`CODE_OF_CONDUCT.md`](https://github.com/featurebyte/featurebyte/blob/main/CODE_OF_CONDUCT.md) and [`CONTRIBUTING.md`](https://github.com/featurebyte/featurebyte/blob/main/CONTRIBUTING.md) guides.
- [ ] I have written tests for the changes made.
- [ ] I have written docstrings in [NumpyDoc format](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard)
- [ ] I have labeled my Pull Request correctly