@jevonyeoh 2023-10-19
nice!! this makes running tests locally so much quicker!